Refactored after Merge, to work again
