+++ WagerMate/Components/Layout/LoginPage.razor
@@ -1,5 +1,7 @@
 @page "/login"
 @using System.ComponentModel.DataAnnotations
+@using WagerMate.Services.auth
+@using WagerMate.Services.user
 
 @layout LoginLayout
 @rendermode InteractiveServer
