+++ WagerMate/Components/Pages/Bet-overview.razor
@@ -3,6 +3,7 @@
 @using System.ComponentModel.DataAnnotations
 @using WagerMate.Services
 @using WagerMate.Data
+@using WagerMate.Services.betting
 @inject IBetService BetService;
 @inject IBetService Service2;
