+++ WagerMate/Components/Pages/Database.razor
@@ -2,6 +2,8 @@
 @using System.ComponentModel.DataAnnotations
 @using WagerMate.Data
 @using WagerMate.Services
+@using WagerMate.Services.betting
+@using WagerMate.Services.user
 @rendermode InteractiveServer
 <PageTitle>Database</PageTitle>
 <h3>Database</h3>
