+++ WagerMate/Components/Pages/Home.razor
@@ -1,5 +1,6 @@
 ﻿@page "/home"
 @using Microsoft.AspNetCore.Components.Authorization
+@using WagerMate.Services.auth
 @inject NavigationManager Navigation
 @inject ICookieService CookieService;
 @rendermode InteractiveServer
