+++ WagerMate/Components/Pages/Userprofile.razor
@@ -1,5 +1,7 @@
 @page "/user"
 @using System.ComponentModel.DataAnnotations
+@using WagerMate.Services.auth
+@using WagerMate.Services.user
 @inject NavigationManager Navigation
 @inject ICookieService CookieService;
 @inject IUserService UserService;
