+++ WagerMate/Service-Implementation/betting/CaseService.cs
@@ -1,7 +1,9 @@
 using WagerMate.Data;
+using WagerMate.Services;
+using WagerMate.Services.betting;
+using WagerMate.Services.database;
 
-
-namespace WagerMate.Services.impl;
+namespace WagerMate.Service_Implementation.betting;
 
 public class CaseService : ICaseService
 {
