+++ WagerMate/Services/betting/IBetService.cs
@@ -1,5 +1,6 @@
 using WagerMate.Data;
-namespace WagerMate.Services;
+
+namespace WagerMate.Services.betting;
 
 public interface IBetService
 {
