+++ WagerMate/Services/betting/ICaseService.cs
@@ -1,6 +1,6 @@
 using WagerMate.Data;
 
-namespace WagerMate.Services;
+namespace WagerMate.Services.betting;
 
 public interface ICaseService
 {
